tests: refactor - move stateful playground tests
