+++ .buildkite/ftr_search_stateful_configs.yml
@@ -5,7 +5,7 @@ disabled:
 defaultQueue: 'n2-4-spot'
 enabled:
   - x-pack/solutions/search/test/functional_search/config.ts
-  - x-pack/solutions/search/test/functional/apps/search_playground/config.ts
+  - x-pack/solutions/search/test/functional_search/config/config.search_playground.ts
   - x-pack/solutions/search/test/functional_search/config/config.feature_flags.ts
   - x-pack/solutions/search/test/api_integration/apis/search_playground/config.ts
   - x-pack/solutions/search/test/functional_solution_sidenav/config.ts
