+++ x-pack/solutions/search/test/functional_search/apps/search_playground/playground_overview.ess.ts
@@ -6,7 +6,7 @@
  */
 
 import { ELASTIC_HTTP_VERSION_HEADER } from '@kbn/core-http-common';
-import { FtrProviderContext } from '../../../functional_search/ftr_provider_context';
+import type { FtrProviderContext } from '../../ftr_provider_context';
 import { createOpenAIConnector } from './utils/create_openai_connector';
 
 import { createLlmProxy, LlmProxy } from './utils/create_llm_proxy';
