+++ x-pack/solutions/search/test/functional_search/apps/search_playground/utils/create_llm_proxy.ts
